upstream_retry when multi-GET not-my-vbucket err
Change-Id: Iebbae35b0509cc16cbaf4610a3e433b60b71b94b

In this implementation of handling not-my-bucket errors during
ascii-to-binary proxying, we use the binary header's opaque field.

moxi uses the opaque field to store either a vbucket id or a key
index, depending on the command opcode.  The key index is for a char
array, and is used to retreive the key for a given command.  Given the
key, moxi can then re-retrieve the vbucket id.

Given the vbucket id, finally, during a not-my-vbucket error, moxi can
then call the incorrect_master() API of libvbucket.