hooked up more tests to 'make test'
moxi_mock.pl script is now reusable, at the cost of some (perl
inspired) readability.  It's now invoked to also run the mock auth,
multitenancy, and multitenancy_default tests scripts (combinations of
spawning a fake rest/http server, a moxi with the right params, and
the right python test scripts to drive it).

Change-Id: If7c16c9709ab59320d804909f578c7f9966a7a29
Reviewed-on: http://review.northscale.com/2644
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>