MB-2689 - add timeout to moxi SASL auth & bucket select requests
In this change, moxi SASl auth and bucket select are still
synchronous, but have additional configurable timeouts (via the
downstream_timeout configuration parameter).

This is done by adding a select() with timeout mcs_io_read(),
which shouldn't be too horrible as SASL auth should be very
infrequent due to long-lived, reused connections.

The alternative solution of rewriting SASL auth and bucket-select, in
comparison, to be asynchronous is a lot of effort.

Change-Id: I61f31fb3a4c4a994ff79d6a9f909ca578ae02236
Reviewed-on: http://review.membase.org/3604
Reviewed-by: Sean Lynch <seanl@literati.org>
Tested-by: Sean Lynch <seanl@literati.org>