Moxi change for enabling event on downstream connection in pause state
when we put the downstream connection in conn_pause state, we disable the
events on that.  So if the connection get closed, we know only when we linked
that connection to upstream connection ,forward the upstream command to that
connection and wait for read event on that.When we wait for read event,
we come to know that connection has closed.
This fix is to enable read event on the downstream connection in pause state,
so  we will know if server send extra data for a command or closes the connection.

Change-Id: Iff50b252a04a8036bb901849953eac557ef8cf9d
Reviewed-on: http://review.couchbase.org/9717
Tested-by: Steve Yen <steve.yen@gmail.com>
Reviewed-by: Steve Yen <steve.yen@gmail.com>