MB-24509: Prevent truncation of documents set with noreply
A reference to the item was not retained for the time the downstream
connection needs the data from the item. The upstream could then free
the item midway through the downstream sending the data to Couchbase
Server.

This issue can be easily observed from a memcache client using the ASCII
protocol:

* Set+noreply a given key to a large (~20MB) document
* Get the document - should match the set document
* Set+noreply to the same key, again with a large document
* Get it a second time - document is now likely to have a small chunk
  of the correct data (in testing ~131kb) and the rest is null bytes,
  padding the document to the correct size

Change-Id: I6a316114aca3ee522d63f08bab17170f56a9d9e2
Reviewed-on: http://review.couchbase.org/78763
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>